Add @shopify/slate-translations plugin
