+++ packages/slate-translations/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@shopify/slate-translations",
+  "version": "1.0.0-beta.16",
+  "main": "index.js",
+  "license": "MIT",
+  "dependencies": {
+    "fs-extra": "^7.0.1",
+    "lodash": "^4.17.11",
+    "path": "^0.12.7"
+  }
+}
